backend/requirements.txt
fastapi
uvicorn[standard]
pyodbc
httpx
python-dotenv
pyyaml
sqlparse
sentence-transformers>=2.2.0
numpy>=1.24.0
